Add CSTG example website
The example website is a static HTML page served by NGINX, that shows
how to use the new CSTG functionality.

In development, the JS SDK is served by webpack-dev-server.

Running the "Launch CSTG (Chrome)" configuration in VS Code will start
webpack-dev-server, as well as a Docker container serving the example
website. webpack-dev-server is configured to allow for hot reload,
as well as debugging the SDK in VS Code.